updated configs and script for esd data processing

+++ src/parallel_dataset.py
@@ -26,10 +26,11 @@ def copy_wavs(cfg: Dict, speaker: int, emotion: str, emotion_id: int, part: str)
         speaker_id = int(speaker_id)
         file_id = int(file_id)
         new_path = f"{cfg.target_directory_path}/wavs/{speaker_id}_{file_id}_{emotion_id}.wav"
-        if os.path.exists(path):
+        txt_path = f"{cfg.target_directory_path}/wavs/{speaker_id}_{file_id}_{emotion_id}.txt"
+        if os.path.exists(path) and os.path.exists(txt_path):
             shutil.copyfile(path, new_path)
-        if part == "test":
-            val_ids_list.append(f"{speaker_id}_{file_id}_{emotion_id}")
+            if part == "test":
+                val_ids_list.append(f"{speaker_id}_{file_id}_{emotion_id}")
     return val_ids_list
 
 
@@ -38,7 +39,7 @@ def build_dataset(cfg):
     os.makedirs(cfg.target_directory_path, exist_ok=True)
     os.makedirs(f"{cfg.target_directory_path}/wavs", exist_ok=True)
 
-    speakers = os.listdir(cfg.source_data_directory)[9:20]
+    speakers = os.listdir(cfg.source_data_directory)[10:21]
     speaker_encodings = cfg.speaker_encodings
     emotions_dict = dict(zip(cfg.emotions, cfg.emotion_ids))
 
@@ -50,13 +51,19 @@ def build_dataset(cfg):
                                        encoding=speaker_encodings[i]).readlines()
 
             for speaker_txt in speaker_transcripts:
-                filename, text, emotion = speaker_txt.split("\t")
-                speaker_id, file_id = filename.split("_")
-                emotion_id = emotions_dict[emotion[:-1]]
-                speaker_id = int(speaker_id)
-                file_id = int(file_id)
-                with open(f"{cfg.target_directory_path}/wavs/{speaker_id}_{file_id}_{emotion_id}.txt", "w") as f:
-                    f.write(text)
+                try:
+                    filename, text, emotion = speaker_txt.split("\t")
+                    speaker_id, file_id = filename.split("_")
+                    emotion = emotion[:-1]
+                    if emotion[-1] == " ":
+                        emotion = emotion[:-1]
+                    emotion_id = emotions_dict[emotion]
+                    speaker_id = int(speaker_id)
+                    file_id = int(file_id)
+                    with open(f"{cfg.target_directory_path}/wavs/{speaker_id}_{file_id}_{emotion_id}.txt", "w") as f:
+                        f.write(text)
+                except ValueError:
+                    pass
 
             for emotion in cfg.emotions:
                 emotion_id = emotions_dict[emotion]
